+++ src/classes/managers/tower.manager.js
@@ -8,7 +8,7 @@
 //====================================================================================================================
 //====================================================================================================================
 
-import Tower from "./tower.model";
+import Tower from "../models/tower.model.js";
 
 class TowerManager {
     static instance = null;
