+++ src/classes/models/tower.model.js
@@ -18,8 +18,8 @@
 //====================================================================================================================
 //====================================================================================================================
 
-import { distance } from '../utils/mathHelper.js';
-import { testLog } from './../utils/testLogger.js';
+import { testLog } from '../../utils/testLogger.js';
+import { distance } from './../../utils/mathHelper.js';
 
 class Tower {
     constructor(userId, towerId, x, y) {
